🛠️refactor: 폴더 변경 및 API 주소 변경
