+++ src/app/(Web)/wedding/invitations/produce/[id]/_constants/index.tsx
@@ -1,1 +1,2 @@
 export { default as DEFAULT_VALUE } from './DefaultValue';
+export * from './DefaultValue';
